Improved code coverage

+++ src/Portfolio.Tracker.Core/Models/AlphaVantageResponse.cs
@@ -9,6 +9,6 @@ namespace Portfolio.Tracker.Core.Models
         public decimal High { get; set; }
         public decimal Low { get; set; }
         public decimal Close { get; set; }
-        public decimal Volume { get; set; }
+        public int Volume { get; set; }
     }
 }
